requirements.txt
# Aristotle API連携に必要なPythonパッケージ
requests>=2.31.0
python-dotenv>=1.0.0

